feat(bot-manager): trade queue (#164)
With the trade queue, you can create, accept, decline / cancel, and
confirm offers reliably. If a job encounters an error, such as the bot
not being available, or Steam gives errors, then the job will be
retried.

Retries can be configured using the following options:
* `strategy` - `"exponential"`, `"linear"` or `"fixed"`.
* `maxTime` - Maximum job age in milliseconds before being deleted.
* `delay` - Delay factor in milliseconds for the different strategies.
* `maxDelay` - Maximum delay between each retry.

Jobs can also be prioritized using `priority`. The highest priority is
1.