feat(bot-manager): queue inventory loads
Instead of getting the inventory from the bot if it is not in the cache, you instead add the
inventory to the queue and it will eventually be loaded depending on the configured retry options.

BREAKING CHANGE: In order to load an inventory, you need to add it to the queue and wait for it to
be fetched.